To split services as controller and finished formol without price

+++ Shared/Entities/Formol.cs
@@ -6,7 +6,7 @@ public class Formol : BourceBase
 {
     public int TimeFormol { get; set; }
 
-    public int HajmFormol { get; set; }
+    public double HajmFormol { get; set; }
 
     public int MultiStage { get; set; }
